Fix colour of spinner to match text
The colour was previously set to be the accent colour, but this is because our site has --button-text-color-hover set to the accent colour. It works less well on Croydon (and others). This change makes the spinner bulk match the text.

There is a problem here, as the rest of the circle is a 10% opacity version of this colour, but we can't achieve that with CSS variables until CSS values and units 5 is supported (to allow for extracting the r,g,b tuple from the hex and reinserting into an rgba). As such, our CSS currently sets this a 10% red, but this is an artifact of Essex. We can achieve a similar (but aliased) effect by setting it to the colour but 0.1px width, but to my eyes this looks more wrong than the circle being the wrong colour given how faint it is. Personally, I don't think the red is sufficiently visable to be recognisable, but I'd take opinions on this.